LISTS: merge from branch 'list'
'list_t' is a generic list implementation.
As for 'l_list_t' it is slower than a classic C array
but it comes with more built-in methods and is safer to use.

We also renamed the old 'list_t' to 'l_list_t' and changed
the related methods' names accordingly.

Signed-off-by: DUBOIN Léo <leo.duboin@epita.fr>